aws/sagemaker_train/requirements.txt
transformers==4.44.2
datasets==2.21.0
pytorch-crf==0.7.2
nltk==3.9.1
scikit-learn==1.5.2
boto3>=1.34
